Refactor rename users_questionnaire -> users_questionnaire_onboarding
