--- androidHyperskillApp/src/main/java/org/hyperskill/app/android/users_questionnaire/onboarding/ui/UsersQuestionnaireOnboardingDefaults.kt
@@ -1,13 +1,0 @@
-package org.hyperskill.app.android.users_questionnaire.onboarding.ui
-
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.ui.unit.dp
-
-object UsersQuestionnaireOnboardingDefaults {
-    val ContentPadding: PaddingValues = PaddingValues(
-        top = 20.dp,
-        start = 20.dp,
-        end = 20.dp,
-        bottom = 8.dp
-    )
-}
