--- shared/src/androidMain/kotlin/org/hyperskill/app/users_questionnaire/onboarding/injection/PlatformUsersQuestionnaireOnboardingComponent.kt
@@ -1,7 +1,0 @@
-package org.hyperskill.app.users_questionnaire.onboarding.injection
-
-import org.hyperskill.app.core.injection.ReduxViewModelFactory
-
-interface PlatformUsersQuestionnaireOnboardingComponent {
-    val reduxViewModelFactory: ReduxViewModelFactory
-}
