+++ shared/src/androidMain/kotlin/org/hyperskill/app/users_questionnaire_onboarding/onboarding/injection/PlatformUsersQuestionnaireOnboardingComponent.kt
@@ -1,0 +1,7 @@
+package org.hyperskill.app.users_questionnaire_onboarding.onboarding.injection
+
+import org.hyperskill.app.core.injection.ReduxViewModelFactory
+
+interface PlatformUsersQuestionnaireOnboardingComponent {
+    val reduxViewModelFactory: ReduxViewModelFactory
+}
