--- shared/src/commonMain/kotlin/org/hyperskill/app/users_questionnaire/onboarding/domain/analytic/UsersQuestionnaireOnboardingViewedHyperskillAnalyticEvent.kt
@@ -1,23 +1,0 @@
-package org.hyperskill.app.users_questionnaire.onboarding.domain.analytic
-
-import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticAction
-import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticEvent
-import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticRoute
-
-/**
- * Represents a view analytic event.
- *
- * JSON payload:
- * ```
- * {
- *     "route": "/onboarding/questionnaire",
- *     "action": "view"
- * }
- * ```
- *
- * @see HyperskillAnalyticEvent
- */
-object UsersQuestionnaireOnboardingViewedHyperskillAnalyticEvent : HyperskillAnalyticEvent(
-    HyperskillAnalyticRoute.Onboarding.UsersQuestionnaire,
-    HyperskillAnalyticAction.VIEW
-)
