--- shared/src/commonMain/kotlin/org/hyperskill/app/users_questionnaire/onboarding/injection/UsersQuestionnaireOnboardingComponentImpl.kt
@@ -1,20 +1,0 @@
-package org.hyperskill.app.users_questionnaire.onboarding.injection
-
-import org.hyperskill.app.core.injection.AppGraph
-import org.hyperskill.app.users_questionnaire.onboarding.presentation.UsersQuestionnaireOnboardingFeature.Action
-import org.hyperskill.app.users_questionnaire.onboarding.presentation.UsersQuestionnaireOnboardingFeature.Message
-import org.hyperskill.app.users_questionnaire.onboarding.presentation.UsersQuestionnaireOnboardingFeature.ViewState
-import ru.nobird.app.presentation.redux.feature.Feature
-
-internal class UsersQuestionnaireOnboardingComponentImpl(
-    private val appGraph: AppGraph
-) : UsersQuestionnaireOnboardingComponent {
-    override val usersQuestionnaireOnboardingFeature: Feature<ViewState, Message, Action>
-        get() = UsersQuestionnaireOnboardingFeatureBuilder.build(
-            analyticInteractor = appGraph.analyticComponent.analyticInteractor,
-            buildVariant = appGraph.commonComponent.buildKonfig.buildVariant,
-            logger = appGraph.loggerComponent.logger,
-            platform = appGraph.commonComponent.platform,
-            resourceProvider = appGraph.commonComponent.resourceProvider
-        )
-}
