+++ shared/src/commonMain/kotlin/org/hyperskill/app/users_questionnaire_onboarding/domain/analytic/UsersQuestionnaireOnboardingAnalyticParams.kt
@@ -1,0 +1,6 @@
+package org.hyperskill.app.users_questionnaire_onboarding.domain.analytic
+
+internal object UsersQuestionnaireOnboardingAnalyticParams {
+    const val PARAM_SOURCE = "source"
+    const val PARAM_INPUT = "input"
+}
