+++ shared/src/commonMain/kotlin/org/hyperskill/app/users_questionnaire_onboarding/injection/UsersQuestionnaireOnboardingComponent.kt
@@ -1,0 +1,10 @@
+package org.hyperskill.app.users_questionnaire_onboarding.injection
+
+import org.hyperskill.app.users_questionnaire_onboarding.presentation.UsersQuestionnaireOnboardingFeature.Action
+import org.hyperskill.app.users_questionnaire_onboarding.presentation.UsersQuestionnaireOnboardingFeature.Message
+import org.hyperskill.app.users_questionnaire_onboarding.presentation.UsersQuestionnaireOnboardingFeature.ViewState
+import ru.nobird.app.presentation.redux.feature.Feature
+
+interface UsersQuestionnaireOnboardingComponent {
+    val usersQuestionnaireOnboardingFeature: Feature<ViewState, Message, Action>
+}
